% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_SE.R
\name{add_SE}
\alias{add_SE}
\title{Add standard error for a fixed parameter.}
\usage{
add_SE(fixed.parameters = NULL, parameters = NULL, SE = NULL)
}
\arguments{
\item{fixed.parameters}{Set of fixed parameters}

\item{parameters}{Set of current parameters}

\item{SE}{Standard error value to be added}
}
\value{
The parameters set with the new SE value
}
\description{
This function is used to add standard error for a fixed parameter.
}
\details{
add_SE adds standard error for a fixed parameter.
}
\examples{
library(phenology)
# Generate a set of fixed parameter: Flat and Min
pfixed<-c(Flat=0, Min=0)
# Add SE for the Flat parameter
pfixed<-add_SE(fixed.parameters=pfixed, parameters="Flat", SE=5)
}
\seealso{
Other Phenology model:
\code{\link[=AutoFitPhenology]{AutoFitPhenology()}},
\code{\link[=BE_to_LBLE]{BE_to_LBLE()}},
\code{\link{Gratiot}},
\code{\link[=LBLE_to_BE]{LBLE_to_BE()}},
\code{\link[=LBLE_to_L]{LBLE_to_L()}},
\code{\link[=L_to_LBLE]{L_to_LBLE()}},
\code{\link{MarineTurtles_2002}},
\code{\link[=MinBMinE_to_Min]{MinBMinE_to_Min()}},
\code{\link[=adapt_parameters]{adapt_parameters()}},
\code{\link[=add_phenology]{add_phenology()}},
\code{\link[=extract_result]{extract_result()}},
\code{\link[=fit_phenology]{fit_phenology()}},
\code{\link[=likelihood_phenology]{likelihood_phenology()}},
\code{\link[=logLik.phenology]{logLik.phenology()}},
\code{\link{map_Gratiot}},
\code{\link[=map_phenology]{map_phenology()}},
\code{\link[=par_init]{par_init()}},
\code{\link[=phenology]{phenology()}},
\code{\link[=phenology2fitRMU]{phenology2fitRMU()}},
\code{\link[=phenology_MHmcmc]{phenology_MHmcmc()}},
\code{\link[=phenology_MHmcmc_p]{phenology_MHmcmc_p()}},
\code{\link[=plot.phenology]{plot.phenology()}},
\code{\link[=plot.phenologydata]{plot.phenologydata()}},
\code{\link[=plot.phenologymap]{plot.phenologymap()}},
\code{\link[=plot_delta]{plot_delta()}},
\code{\link[=plot_phi]{plot_phi()}},
\code{\link[=print.phenology]{print.phenology()}},
\code{\link[=print.phenologymap]{print.phenologymap()}},
\code{\link[=print.phenologyout]{print.phenologyout()}},
\code{\link[=remove_site]{remove_site()}},
\code{\link{result_Gratiot}},
\code{\link{result_Gratiot1}},
\code{\link{result_Gratiot2}},
\code{\link{result_Gratiot_Flat}},
\code{\link[=summary.phenology]{summary.phenology()}},
\code{\link[=summary.phenologymap]{summary.phenologymap()}},
\code{\link[=summary.phenologyout]{summary.phenologyout()}}
}
\author{
Marc Girondot
}
\concept{Phenology model}
