% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean.knitr}
\alias{clean.knitr}
\title{Delete temporary files created during knitr compile}
\usage{
clean.knitr()
}
\value{
Nothing
}
\description{
Delete temporary files created during knitr compile in working directory
}
\details{
clean.knitr deletes temporary files created during knitr compile
}
\examples{
\dontrun{
clean.knitr()
}
}
\author{
Marc Girondot
}

