% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenology.R
\name{.read_phenology}
\alias{.read_phenology}
\title{The function .read_phenology}
\usage{
.read_phenology(obj_list = NULL, header = NULL, reference = NULL,
  month_ref = NULL, format = NULL, nm = NULL, sep.dates = NULL,
  silent = FALSE)
}
\arguments{
\item{obj_list}{A list with different datasets}

\item{header}{Does the timeseries has header.}

\item{reference}{Date used as reference. Is the day 1.}

\item{month_ref}{Reference month. Generally will be 1 (January) or 7 (July).}

\item{format}{Format for dates.}

\item{nm}{Path for file or name}

\item{sep.dates}{Character used as date separator}

\item{silent}{Should warning b shown}
}
\value{
Return a dataframe
}
\description{
Function for package Phenology
}
\details{
.read_phenology translate file format
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
