\name{growlnotify}
\alias{growlnotify}
\title{Send growl notification for MacOs X system.}
\usage{
  growlnotify(textinfo = "", help = FALSE)
}
\arguments{
  \item{textinfo}{Text to display in the growlnotify
  window}

  \item{help}{If TRUE, an help is displayed}
}
\value{
  None
}
\description{
  This function is used to send a notification to Mac user.
}
\details{
  growlnotify send growl notification for MacOs X systems.
}
\author{
  Marc Girondot
}

