% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/insidesearch.R
\name{insidesearch}
\alias{insidesearch}
\title{Search a string within files of a folder
#'}
\usage{
insidesearch(path = ".", pattern = "*.R", ...,
  search = stop("A text to be searched for is necessary"))
}
\arguments{
\item{path}{Path of the folder to search}

\item{pattern}{Pattern for file names to searcg}

\item{...}{Options for readLines(), example warn = FALSE}

\item{search}{String to search in files}
}
\value{
Return a vector with filenames
}
\description{
Search for a string inside the files of a folder \cr
}
\details{
insidesearch Search a string within files of a folder
}
\examples{
\dontrun{
library(phenology)
# Read a file with data
}
}
\author{
Marc Girondot
}

