% Generated by roxygen2 (4.0.1): do not edit by hand
\name{local.search}
\alias{local.search}
\title{Return path of file searched for in local disk based on its file name}
\usage{
local.search(pattern, directory = "", folder = "$HOME")
}
\arguments{
\item{pattern}{The name of file to be searched for. Can use wildcards *}

\item{directory}{The path of directory to be explored in for Windows}

\item{folder}{The path of folder to be explored in for Unix based systems}
}
\value{
A vector with paths
}
\description{
Return path of file searched for in local disk based on its file name.
It has been tested only with Windows XP and MacOSX.
}
\details{
local.search() returns path of file serached in local disk based on its file name
}
\examples{
\dontrun{
RnwFiles <- local.search("*.Rnw")
}
}
\author{
Marc Girondot
}

