% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outLR.R
\docType{data}
\name{outLR}
\alias{outLR}
\title{Database of leatherback CMR in French Guiana}
\format{
A dataframe with database of leatherback CMR in French Guiana
}
\usage{
outLR
}
\description{
Database of leatherback CMR in French Guiana
}
\details{
Database of leatherback CMR in French Guiana
}
\examples{
\dontrun{
library(phenology)
# Read a file with result
data(outLR)
data_f_21 <- Tagloss_format(outLR, model="21")
}
}
\references{
Rivalan, P., Godfrey, M.H., Prévot-Julliard, A.-C., 
Girondot, M., 2005. Maximum likelihood estimates of tag loss in 
leatherback sea turtles. Journal of Wildlife Management 69, 540-548.
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\keyword{datasets}
