% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{phenology}
\alias{phenology}
\title{Run a shiny application for basic functions of phenology package}
\usage{
phenology()
}
\value{
Nothing
}
\description{
Run a shiny application for basic functions of phenology package.
Thanks to Adriana Cortés Gomés and Joana Hancock for their help with translation.
}
\details{
phenology runs a shiny application for basic functions of phenology package
}
\examples{
\dontrun{
library(phenology)
phenology()
}
}
\seealso{
Other Phenology model:
\code{\link[=AutoFitPhenology]{AutoFitPhenology()}},
\code{\link[=BE_to_LBLE]{BE_to_LBLE()}},
\code{\link{Gratiot}},
\code{\link[=LBLE_to_BE]{LBLE_to_BE()}},
\code{\link[=LBLE_to_L]{LBLE_to_L()}},
\code{\link[=L_to_LBLE]{L_to_LBLE()}},
\code{\link{MarineTurtles_2002}},
\code{\link[=MinBMinE_to_Min]{MinBMinE_to_Min()}},
\code{\link[=adapt_parameters]{adapt_parameters()}},
\code{\link[=add_SE]{add_SE()}},
\code{\link[=add_phenology]{add_phenology()}},
\code{\link[=extract_result]{extract_result()}},
\code{\link[=fit_phenology]{fit_phenology()}},
\code{\link[=likelihood_phenology]{likelihood_phenology()}},
\code{\link[=logLik.phenology]{logLik.phenology()}},
\code{\link{map_Gratiot}},
\code{\link[=map_phenology]{map_phenology()}},
\code{\link[=par_init]{par_init()}},
\code{\link[=phenology2fitRMU]{phenology2fitRMU()}},
\code{\link[=phenology_MHmcmc]{phenology_MHmcmc()}},
\code{\link[=phenology_MHmcmc_p]{phenology_MHmcmc_p()}},
\code{\link[=plot.phenology]{plot.phenology()}},
\code{\link[=plot.phenologydata]{plot.phenologydata()}},
\code{\link[=plot.phenologymap]{plot.phenologymap()}},
\code{\link[=plot_delta]{plot_delta()}},
\code{\link[=plot_phi]{plot_phi()}},
\code{\link[=print.phenology]{print.phenology()}},
\code{\link[=print.phenologymap]{print.phenologymap()}},
\code{\link[=print.phenologyout]{print.phenologyout()}},
\code{\link[=remove_site]{remove_site()}},
\code{\link{result_Gratiot}},
\code{\link{result_Gratiot1}},
\code{\link{result_Gratiot2}},
\code{\link{result_Gratiot_Flat}},
\code{\link[=summary.phenology]{summary.phenology()}},
\code{\link[=summary.phenologymap]{summary.phenologymap()}},
\code{\link[=summary.phenologyout]{summary.phenologyout()}}
}
\author{
Marc Girondot
}
\concept{Phenology model}
