\name{phenology_swot}
\alias{phenology_swot}
\title{Fit a nesting season of marine turtles}
\usage{
  phenology_swot(header = NULL, reference = NULL,
    format = NULL)
}
\arguments{
  \item{header}{Does the timeseries has header.}

  \item{reference}{Date used as reference - day 1.}

  \item{format}{Format for dates.}
}
\value{
  None
}
\description{
  This function tries to make very simple the use of this
  package.\cr In most of the cases, values for header,
  reference and format need not to be set\cr because they
  are detected automically.
}
\details{
  phenology_swot is a simplified function for phenology
  package.
}
\author{
  Marc Girondot \email{marc.girondot@u-psud.fr}
}

