% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phenologydata.R
\name{plot.phenologydata}
\alias{plot.phenologydata}
\title{Plot the phenology from a add_phenology().}
\usage{
\method{plot}{phenologydata}(
  x,
  ...,
  series = "all",
  moon = FALSE,
  season = NULL,
  col.observations = "black",
  col.minimum.observations = "blue",
  col.grouped.observations = "green"
)
}
\arguments{
\item{x}{A result file generated by add_phenology}

\item{...}{Parameters used by plot}

\item{series}{Name or number of series to be plotted or 'all'}

\item{moon}{If TRUE, the moon phase is ploted. Default is FALSE}

\item{season}{Which season to plot}

\item{col.observations}{Color of the points}

\item{col.minimum.observations}{Color of the points indicating minimum counts}

\item{col.grouped.observations}{Color of the lines indicating grouped observations}
}
\value{
Nothing.
}
\description{
The function plot.phenologydata plots the observations.\cr
}
\details{
plot.phenologydata plots the observation data of phenology.
}
\examples{
\dontrun{
library(phenology)
# Read a file with data
data(Gratiot)
# Generate a formatted list nammed data_Gratiot 
data_Gratiot <- add_phenology(Gratiot, name = "Complete", 
		reference = as.Date("2001-01-01"), format="\%d/\%m/\%Y")

plot(data_Gratiot)
}
}
\seealso{
Other Phenology model:
\code{\link[=AutoFitPhenology]{AutoFitPhenology()}},
\code{\link[=BE_to_LBLE]{BE_to_LBLE()}},
\code{\link{Gratiot}},
\code{\link[=LBLE_to_BE]{LBLE_to_BE()}},
\code{\link[=LBLE_to_L]{LBLE_to_L()}},
\code{\link[=L_to_LBLE]{L_to_LBLE()}},
\code{\link{MarineTurtles_2002}},
\code{\link[=MinBMinE_to_Min]{MinBMinE_to_Min()}},
\code{\link[=adapt_parameters]{adapt_parameters()}},
\code{\link[=add_SE]{add_SE()}},
\code{\link[=add_phenology]{add_phenology()}},
\code{\link[=extract_result]{extract_result()}},
\code{\link[=fit_phenology]{fit_phenology()}},
\code{\link[=likelihood_phenology]{likelihood_phenology()}},
\code{\link[=logLik.phenology]{logLik.phenology()}},
\code{\link{map_Gratiot}},
\code{\link[=map_phenology]{map_phenology()}},
\code{\link[=par_init]{par_init()}},
\code{\link[=phenology]{phenology()}},
\code{\link[=phenology2fitRMU]{phenology2fitRMU()}},
\code{\link[=phenology_MHmcmc]{phenology_MHmcmc()}},
\code{\link[=phenology_MHmcmc_p]{phenology_MHmcmc_p()}},
\code{\link[=plot.phenology]{plot.phenology()}},
\code{\link[=plot.phenologymap]{plot.phenologymap()}},
\code{\link[=plot_delta]{plot_delta()}},
\code{\link[=plot_phi]{plot_phi()}},
\code{\link[=print.phenology]{print.phenology()}},
\code{\link[=print.phenologymap]{print.phenologymap()}},
\code{\link[=print.phenologyout]{print.phenologyout()}},
\code{\link[=remove_site]{remove_site()}},
\code{\link{result_Gratiot}},
\code{\link{result_Gratiot1}},
\code{\link{result_Gratiot2}},
\code{\link{result_Gratiot_Flat}},
\code{\link[=summary.phenology]{summary.phenology()}},
\code{\link[=summary.phenologymap]{summary.phenologymap()}},
\code{\link[=summary.phenologyout]{summary.phenologyout()}}
}
\author{
Marc Girondot \email{marc.girondot@gmail.com}
}
\concept{Phenology model}
