\name{shift_sinusoid}
\alias{shift_sinusoid}
\title{Shift sinusoid information.}
\usage{
  shift_sinusoid(parameters = NULL, from = "", to = "1",
    help = FALSE)
}
\arguments{
  \item{parameters}{set of parameters}

  \item{from}{The number of series to change}

  \item{to}{The number of series to change}

  \item{help}{If TRUE, an help is displayed}
}
\value{
  Return a set of modified parameters
}
\description{
  This function is used to shift sinusoid parameters from
  '', '1' or '2'.
}
\details{
  shift_sinusoid shift sinusoid information.
}
\author{
  Marc Girondot
}

