\name{add_SD}
\alias{add_SD}
\title{Add SD for a fixed parameter.}
\usage{
  add_SD(parametersfixed = NULL, parameter = NULL,
    SD = NULL, help = FALSE)
}
\arguments{
  \item{parametersfixed}{Set of fixed parameters}

  \item{parameter}{Set of current parameters}

  \item{SD}{Standard deviation value to be added}

  \item{help}{If TRUE, an help is displayed}
}
\value{
  The parameters set with the new SD value
}
\description{
  This function is used to add standard deviation for a
  fixed parameter.
}
\details{
  add_SD adds SD for a fixed parameter.
}
\author{
  Marc Girondot
}

