% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.par.mcmcComposite}
\alias{as.par.mcmcComposite}
\title{Extract parameters at maximum likelihood from the result of MHmcmc}
\usage{
as.par.mcmcComposite(x)
}
\arguments{
\item{x}{A result from MHmcmc search}
}
\value{
A vector with parameters at maximum likelihood
}
\description{
Take a mcmcComposite object and create a vector object with parameter value at maximum likelihood.
It also says at which iteration the maximum lihelihood has been observed.
}
\details{
as.par.mcmcComposite Extract parameters at maximum likelihood from the result of MHmcmc
}
\examples{
\dontrun{
library(phenology)
data(result_Gratiot_mcmc)
x <- as.par.mcmcComposite(result_Gratiot_mcmc)
}
}
\author{
Marc Girondot
}

