\name{extract_result}
\alias{extract_result}
\title{Extract the set of parameters from a result object.}
\usage{
  extract_result(result = res, help = FALSE)
}
\arguments{
  \item{result}{A result file}

  \item{help}{If TRUE, an help is displayed}
}
\value{
  Return the set of fitted parameters
}
\description{
  The function "extract_result" permits to extract the set
  of parameters from a result object obtained after
  fit_phenology.
}
\details{
  extract_result get the fitted parameters from a result
  object.
}
\author{
  Marc Girondot
}

