% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.phenology.R
\name{logLik.phenology}
\alias{logLik.phenology}
\title{Return Log Likelihood of a fit generated by fit_phenology}
\usage{
\method{logLik}{phenology}(object, ...)
}
\arguments{
\item{object}{A result file generated by fit_phenology}

\item{...}{Not used}
}
\value{
The Log Likelihood value of the fitted model and data
}
\description{
Return Log Likelihood of a fit generated by fit_phenology
}
\details{
logLik.phenology Return Log Likelihood of a fit
}
\examples{
\dontrun{
library(phenology)
data(result_Gratiot)
logLik(result_Gratiot)
AIC(result_Gratiot)
}
}
\seealso{
Other Phenology model: 
\code{\link{AutoFitPhenology}()},
\code{\link{BE_to_LBLE}()},
\code{\link{Gratiot}},
\code{\link{LBLE_to_BE}()},
\code{\link{LBLE_to_L}()},
\code{\link{L_to_LBLE}()},
\code{\link{MarineTurtles_2002}},
\code{\link{MinBMinE_to_Min}()},
\code{\link{adapt_parameters}()},
\code{\link{add_SE}()},
\code{\link{add_phenology}()},
\code{\link{extract_result}()},
\code{\link{fit_phenology}()},
\code{\link{likelihood_phenology}()},
\code{\link{map_Gratiot}},
\code{\link{map_phenology}()},
\code{\link{par_init}()},
\code{\link{phenology2fitRMU}()},
\code{\link{phenology_MHmcmc_p}()},
\code{\link{phenology_MHmcmc}()},
\code{\link{phenology}()},
\code{\link{plot.phenologymap}()},
\code{\link{plot.phenology}()},
\code{\link{plot_delta}()},
\code{\link{plot_phi}()},
\code{\link{print.phenologymap}()},
\code{\link{print.phenologyout}()},
\code{\link{print.phenology}()},
\code{\link{remove_site}()},
\code{\link{result_Gratiot1}},
\code{\link{result_Gratiot2}},
\code{\link{result_Gratiot_Flat}},
\code{\link{result_Gratiot_mcmc}},
\code{\link{result_Gratiot}},
\code{\link{summary.phenologymap}()},
\code{\link{summary.phenologyout}()},
\code{\link{summary.phenology}()}
}
\author{
Marc Girondot
}
\concept{Phenology model}
