% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logLik.phenology}
\alias{logLik.phenology}
\title{Return Log Likelihood of a fit generated by fit_phenology}
\usage{
\method{logLik}{phenology}(object, ...)
}
\arguments{
\item{object}{A result file generated by fit_phenology}

\item{...}{Not used}
}
\value{
The Log Likelihood value of the fitted model and data
}
\description{
Return Log Likelihood of a fit generated by fit_phenology
}
\details{
logLik.phenology Return Log Likelihood of a fit
}
\examples{
\dontrun{
library(phenology)
data(result_Gratiot)
logLik(result_Gratiot)
AIC(result_Gratiot)
}
}
\author{
Marc Girondot
}

