\docType{data}
\name{map_Gratiot}
\alias{map_Gratiot}
\title{Likelihood map of Leatherback nest counts}
\format{A list with Gratiot data and the result of the fit.}
\usage{
map_Gratiot
}
\description{
Likelihood map of Leatherback nest counts from Gratiot et
al. (2006) Figure 1. A intraseasonal periodic pattern was
searched for varying Phi and Delta parameters.
}
\details{
Likelihood map of Leatherback nest counts from Gratiot et
al. (2006) Figure 1
}
\examples{
library(phenology)
# Read a file with likelihood map
data(map_Gratiot)
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\references{
Gratiot, N., Gratiot, J., de Thoisy, B. & Kelle, L. 2006.
Estimation of marine turtles nesting season from incomplete
data ; statistical adjustment of a sinusoidal function.
Animal Conservation, 9, 95-102.
}
\keyword{datasets}

