% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarineTurtles_2002.R
\docType{data}
\name{MarineTurtles_2002}
\alias{MarineTurtles_2002}
\title{Database of tagged marine turtles in 2002}
\format{
data.frame with 2 columns:
\itemize{
  \item \code{Date}: The date the female has been seen on the beach (morning date of the night)
  \item \code{ID}: The unique identifier of the female
}
}
\usage{
MarineTurtles_2002
}
\description{
Extraction of 2002 PIT tagged marine turtles database
}
\details{
Database of tagged marine turtles in 2002
}
\examples{
library(phenology)
data(MarineTurtles_2002)
}
\seealso{
Other Phenology model: 
\code{\link{AutoFitPhenology}()},
\code{\link{BE_to_LBLE}()},
\code{\link{Gratiot}},
\code{\link{LBLE_to_BE}()},
\code{\link{LBLE_to_L}()},
\code{\link{L_to_LBLE}()},
\code{\link{MinBMinE_to_Min}()},
\code{\link{adapt_parameters}()},
\code{\link{add_SE}()},
\code{\link{add_phenology}()},
\code{\link{extract_result}()},
\code{\link{fit_phenology}()},
\code{\link{likelihood_phenology}()},
\code{\link{logLik.phenology}()},
\code{\link{map_Gratiot}},
\code{\link{map_phenology}()},
\code{\link{par_init}()},
\code{\link{phenology2fitRMU}()},
\code{\link{phenology_MHmcmc_p}()},
\code{\link{phenology_MHmcmc}()},
\code{\link{phenology}()},
\code{\link{plot.phenologymap}()},
\code{\link{plot.phenology}()},
\code{\link{plot_delta}()},
\code{\link{plot_phi}()},
\code{\link{print.phenologymap}()},
\code{\link{print.phenologyout}()},
\code{\link{print.phenology}()},
\code{\link{remove_site}()},
\code{\link{result_Gratiot1}},
\code{\link{result_Gratiot2}},
\code{\link{result_Gratiot_Flat}},
\code{\link{result_Gratiot_mcmc}},
\code{\link{result_Gratiot}},
\code{\link{summary.phenologymap}()},
\code{\link{summary.phenologyout}()},
\code{\link{summary.phenology}()}
}
\author{
Marc Girondot
}
\concept{Phenology model}
\keyword{datasets}
