% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarineTurtles_2002.R
\docType{data}
\name{MarineTurtles_2002}
\alias{MarineTurtles_2002}
\title{Database of tagged marine turtles in 2002}
\format{data.frame with 2 columns:
\itemize{
  \item \code{Date}: The date the female has been seen on the beach (morning date of the night)
  \item \code{ID}: The unique identifier of the female
}}
\usage{
MarineTurtles_2002
}
\description{
Extraction of 2002 PIT tagged marine turtles database
}
\details{
Database of tagged marine turtles in 2002
}
\examples{
library(phenology)
data(MarineTurtles_2002)
}
\author{
Marc Girondot
}
\keyword{datasets}
