\docType{package}
\name{phenology-package}
\alias{phenology-package}
\title{The package phenology}
\description{
  A package to fit seasonality counts
}
\details{
  The package phenology fits a parametric function that
  describes phenology

  \tabular{ll}{ Package: \tab phenology\cr Type: \tab
  Package\cr Version: \tab 3.25\cr Date: \tab 2012-04-05\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }
}
\author{
  Marc Girondot \email{marc.girondot@u-psud.fr}
}
\references{
  Girondot, M. 2010. Estimating density of animals during
  migratory waves: application to marine turtles at nesting
  site.  Endangered Species Research, 12, 85-105.

  Girondot, M. 2010. Editorial: The zero counts. Marine
  Turtle Newsletter, 129, 5-6.
}
\keyword{Ecology}
\keyword{Phenology}
\keyword{Seasonality}

