% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phenology.R
\name{phenology}
\alias{phenology}
\title{Run a shiny application for basic functions of phenology}
\usage{
phenology()
}
\value{
Nothing
}
\description{
Run a shiny application for basic functions of phenology
}
\details{
phenology runs a shiny application for basic functions of phenology
}
\examples{
\dontrun{
library(phenology)
phenology()
}
}
\author{
Marc Girondot
}

