\name{plot_phenology}
\alias{plot_phenology}
\title{Plot the phenology from a result.}
\usage{
  plot_phenology(result = NULL, data = NULL,
    parameters = NULL, parametersfixed = NA,
    series = "all", pdf = FALSE, help = FALSE, ...)
}
\arguments{
  \item{result}{A result file generated by fit_phenology}

  \item{data}{A dataset generated by add_format}

  \item{pdf}{TRUE or FALSE, indicates if a pdf file is
  generated.}

  \item{parametersfixed}{Set of fixed parameters}

  \item{parameters}{Set of parameters to be changed}

  \item{series}{Number of series to be analyzed or 'all'}

  \item{...}{Parameters used for graphics}

  \item{help}{If TRUE, an help is displayed}
}
\value{
  Return A map generated with map_phenology
}
\description{
  The function "plot_phenology" plots the phenology graph
  from a result.
}
\details{
  plot_phenology plots the phenology.
}
\author{
  Marc Girondot
}

