\name{plot_phenology}
\alias{plot_phenology}
\title{Plot the phenology from a result.}
\usage{
  plot_phenology(result = NULL, data = NULL,
    parameters = NULL, parametersfixed = NA,
    series = "all", pdf = FALSE, moon = FALSE,
    help = FALSE, ...)
}
\arguments{
  \item{result}{A result file generated by fit_phenology}

  \item{data}{A dataset generated by add_format}

  \item{pdf}{TRUE or FALSE, indicates if a pdf file is
  generated.}

  \item{parametersfixed}{Set of fixed parameters}

  \item{parameters}{Set of parameters to be changed}

  \item{series}{Number of series to be analyzed or 'all'}

  \item{...}{Parameters used for graphics (xlab, ylab,
  ylim, cex and pch)}

  \item{moon}{If TRUE, the moon phase is ploted. Default is
  FALSE}

  \item{help}{If TRUE, an help is displayed}
}
\value{
  Return A list of lists with formated outputs\cr For each
  site:\cr $site=name of the site\cr $estimate1=Estimation
  of counts not taking into account the observations\cr
  $sd1=The SD of estimation1\cr $estimate2=Estimation of
  counts taking into account the observations\cr $sd2=The
  SD of estimation2
}
\description{
  The function "plot_phenology" plots the phenology graph
  from a result.
}
\details{
  plot_phenology plots the phenology.
}
\examples{
library(phenology)
# Read a file with data
# Gratiot<-read.delim("http://max2.ese.u-psud.fr/epc/conservation/BI/Complete.txt", , header=FALSE)
data(Gratiot)
# Generate a formatted list nammed data_Gratiot
data_Gratiot<-add_format(origin=NULL, add=Gratiot, name="Complete", reference=as.Date("2001-01-01"), format="\%d/\%m/\%Y")
# Generate initial points for the optimisation
parg<-par_init(data_Gratiot, parametersfixed=NULL)
# Run the optimisation
# result_Gratiot<-fit_phenology(data=data_Gratiot, parametersfit=parg, parametersfixed=NULL, trace=1)
data(result_Gratiot)
# Plot the phenology and get some stats
output<-plot_phenology(result=result_Gratiot, pdf=FALSE)
}
\author{
  Marc Girondot
}

