\name{plot_phi}
\alias{plot_phi}
\title{Plot the best likelihood for fixed Phi value.}
\usage{
  plot_phi(map = NULL, pdf = FALSE, pdfname = "Map.pdf",
    help = FALSE)
}
\arguments{
  \item{map}{A map generated with map_phenology}

  \item{pdf}{TRUE or FALSE, indicates if a pdf file is
  generated.}

  \item{pdfname}{Name of pdf file}

  \item{help}{If TRUE, an help is displayed}
}
\value{
  Return None
}
\description{
  The function "plot_phi" plots the best likelihood for
  fixed Phi value.
}
\details{
  plot_phi plots the best likelihood for fixed Phi value.
}
\author{
  Marc Girondot
}

