\name{remove_site}
\alias{remove_site}
\title{Removes site information from a set of parameters.}
\usage{
  remove_site(parameters = NULL, help = FALSE)
}
\arguments{
  \item{parameters}{Set of parameters}

  \item{help}{If TRUE, an help is displayed}
}
\value{
  Return a set of modified parameters
}
\description{
  This function is used to remove the information of the
  site from a set of parameters. It can be used to other
  timeseries after.
}
\details{
  remove_site removes beach information from a set of
  parameters.
}
\author{
  Marc Girondot
}

