% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_Gratiot_Flat.R
\docType{data}
\name{result_Gratiot_Flat}
\alias{result_Gratiot_Flat}
\title{Result of the fit of Leatherback nest counts}
\format{A list with Gratiot data and the result of the fit.}
\usage{
result_Gratiot_Flat
}
\description{
Result of the fit of Leatherback nest counts from Gratiot et al. (2006) Figure 1.
The phenology has been fitted with MinE, MinB, Max, LengthB, LengthE, Peak, Theta.
The Flat parameter is set to 0 and is not fitted.
}
\details{
Result of the fit of Leatherback nest counts from Gratiot et al. (2006) Figure 1
}
\examples{
library(phenology)
# Read a file with result
data(result_Gratiot_Flat)
}
\references{
Gratiot, N., Gratiot, J., de Thoisy, B. & Kelle, L. 2006. 
            Estimation of marine turtles nesting season from incomplete 
            data ; statistical adjustment of a sinusoidal function. Animal 
            Conservation, 9, 95-102.
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\keyword{datasets}
